fix(multicloud): address PR review feedback
* deploy.py: guard tagless --aws-image against ValueError crash
  (greptile P1, deploy.py:721)
* deploy.py: check_auth only verifies clouds actually used in config,
  so single-cloud deployments no longer fail on missing credentials
  for the other cloud (greptile P1, deploy.py:262)
* aws/eks/create_cluster.sh: exit 1 when EFS doesn't reach 'available'
  within 150s instead of continuing into a cryptic mount-target error
  (greptile P1, create_cluster.sh:78)
* multicloud/README.md: drop references to BUGS.md, DEPLOY_NOTES.md,
  RECOMMENDATIONS.md (they are scratch files outside the repo)
  (YuanTingHsieh)
* tests: merge test_k8s_launcher.py into k8s_launcher_test.py;
  rename test_job_launcher_utils.py to job_launcher_utils_test.py
  to match project naming convention (YuanTingHsieh, greptile P2)
